+++ jupyter-book/_toc.yml
@@ -12,11 +12,13 @@
       - file: python_scripts/02_numerical_pipeline
       - file: python_scripts/02_numerical_pipeline_ex_01
       - file: python_scripts/02_numerical_pipeline_sol_01
+      - file: python_scripts/02_numerical_pipeline_scaling
   - file: predictive_modeling_pipeline/03_categorical_pipeline_index
     sections:
       - file: python_scripts/03_categorical_pipeline
       - file: python_scripts/03_categorical_pipeline_ex_01
       - file: python_scripts/03_categorical_pipeline_sol_01
+      - file: python_scripts/03_categorical_pipeline_column_transformer
       - file: python_scripts/03_categorical_pipeline_ex_02
       - file: python_scripts/03_categorical_pipeline_sol_02
   - file: predictive_modeling_pipeline/04_parameter_tuning_index
@@ -24,6 +26,7 @@
       - file: python_scripts/04_parameter_tuning
       - file: python_scripts/04_parameter_tuning_ex_01
       - file: python_scripts/04_parameter_tuning_sol_01
+      - file: python_scripts/04_parameter_tuning_search
       - file: python_scripts/04_parameter_tuning_ex_02
       - file: python_scripts/04_parameter_tuning_sol_02
