+++ jupyter-book/predictive_modeling_pipeline/03_categorical_pipeline_index.md
@@ -1,4 +1,4 @@
-# Preprocessing for categorical data
+# Handling categorical data
 
 ```{tableofcontents}
 ```
