+++ jupyter-book/predictive_modeling_pipeline/04_parameter_tuning_index.md
@@ -1,4 +1,4 @@
-# Parameter tuning
+# Hyper-parameter tuning
 
 ```{tableofcontents}
 ```
